Partner kim

+++ src/main/java/com/scit48/common/dto/UserDTO.java
@@ -4,6 +4,7 @@ import com.scit48.common.enums.LanguageLevel;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.scit48.common.enums.Gender;
 import com.scit48.common.domain.entity.UserEntity;
+import com.scit48.common.enums.LanguageLevel;
 import lombok.*;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -15,40 +16,40 @@ import java.time.LocalDateTime;
 @Builder
 @ToString
 public class UserDTO {
-
+	
 	private Long id; // PK
 
 	private String memberId;
 
 	// 중요: 클라이언트 -> 서버 전송(입력) 시에는 동작하지만,
-	// 서버 -> 클라이언트 응답(조회) 시에는 JSON에서 제외됩니다.
+// 서버 -> 클라이언트 응답(조회) 시에는 JSON에서 제외됩니다.
 	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
 	private String password;
-
+	
 	private String nickname;
-
+	
 	private Gender gender; // Enum
-
+	
 	private String intro;
-
+	
 	private Integer age;
-
+	
 	private String nation;
-
-	private BigDecimal manner;
-
+	
+	private double manner;
+	
 	private String profileImageName;
-
+	
 	private String profileImagePath;
-
+	
 	private String nativeLanguage;
 
 	private LanguageLevel levelLanguage;
 
 	private LocalDateTime createdAt;
 
-	// --- DTO 변환 메서드 (편의성 제공) ---
-
+// --- DTO 변환 메서드 (편의성 제공) ---
+	
 	/*
 	 * DTO -> Entity 변환
 	 * (비밀번호 암호화 처리는 Service 레이어에서 수행 후 별도로 주입하거나,
@@ -64,14 +65,14 @@ public class UserDTO {
 				.intro(this.intro)
 				.age(this.age)
 				.nation(this.nation)
-				.manner(this.manner)
+				.manner(this.manner) // 값이 없으면 Entity Builder에서 36.5로 처리됨
 				.nativeLanguage(this.nativeLanguage)
 				.levelLanguage(this.levelLanguage)
 				.profileImageName(this.profileImageName)
 				.profileImagePath(this.profileImagePath)
 				.build();
 	}
-
+	
 	/*
 	 * Entity -> DTO 변환
 	 */
@@ -94,5 +95,6 @@ public class UserDTO {
 				.createdAt(entity.getCreatedAt())
 				.build();
 	}
-
+	
 }
+
